storage: push bulkio limit down to per-chunk
We previously only ratelimited per-file, but since we already have a chunked file writer for fsyncing, we can easily rate limit more granularly.

Release note: none.
